pyproject.toml
[tool.poetry]
name = "cosmo-s"
version = "0.1.0"
description = ""
authors = ["Даниил <samtesred@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
django = "^5.0.6"
djangorestframework = "^3.15.1"
requests = "^2.32.2"
pycryptodome = "^3.20.0"
pydantic = "^2.7.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
